изменил связи. добавил новый скрипт

+++ travelplanner/src/main/java/com/netcracker/travelplanner/repository/EdgeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import com.netcracker.travelplanner.entities.Edge;
 
 @Repository
-public interface EdgeRepository extends JpaRepository<Edge,Long> {
+public interface EdgeRepository extends JpaRepository<Edge,Integer> {
 
 
 }
